Added example of WaterCloud SAR RT model
This commit adds a first very simple version of the Water Cloud SAR
forward model. The model is a simplification of the one we want, but
demonstrates how the code can be written. As it stands, and given the
simplicity of the SAR forward operator, there is no need to emulate it
and its gradient, as they can be easily calculated. I have added some
tests to this too.